shifted patch trigger to main framework class
