+++ plugins/fused-ops-and-kernels/src/fms_acceleration_foak/models/llama.py
@@ -34,102 +34,6 @@ from ..kernels.unsloth.rms_layernorm import fast_rms_layernorm
 from ..kernels.unsloth.rope_embedding import fast_rope_embedding
 from .utils import KEY_MLP, KEY_O, KEY_QKV, build_lora_fused_ops, trigger_fused_ops
 
-# TODO: have a generic version of this rule
-# - do regex on RMSNorm class name
-# - check on the tensors required for fast_rms_layernorm
-RULE_LLAMA_RMS = ModelPatcherRule(
-    rule_id="llama-rms",
-    trigger=ModelPatcherTrigger(check=LlamaRMSNorm),
-    forward=fast_rms_layernorm,
-)
-
-# TODO: have a generic version of this rule
-# - do regex on Attention class name
-# - have a set of qkv / o module names and check on that
-RULE_LLAMA_QKVO = ModelPatcherRule(
-    rule_id="llama-qkvo",
-    trigger=combine_triggers(
-        ModelPatcherTrigger(
-            check=partial(
-                trigger_fused_ops,
-                attn_cls=LlamaAttention,
-                submodule_names=["q_proj", "k_proj", "v_proj"],
-            )
-        ),
-        ModelPatcherTrigger(
-            check=partial(
-                trigger_fused_ops,
-                attn_cls=LlamaAttention,
-                submodule_names=["o_proj"],
-            )
-        ),
-        logic="OR",
-    ),
-    forward_builder=combine_functions(
-        partial(
-            build_lora_fused_ops,
-            submodule_names=["q_proj", "k_proj", "v_proj"],
-            fused_op=KEY_QKV,
-        ),
-        partial(
-            build_lora_fused_ops,
-            submodule_names=["o_proj"],
-            fused_op=KEY_O,
-        ),
-        logic="APPEND",
-    ),
-    forward_builder_args=["base_type"],
-)
-
-RULE_LLAMA_MLP = ModelPatcherRule(
-    rule_id="llama-mlp",
-    trigger=ModelPatcherTrigger(
-        check=partial(
-            trigger_fused_ops,
-            attn_cls=LlamaMLP,
-            submodule_names=["up_proj", "down_proj", "gate_proj"],
-        )
-    ),
-    forward_builder=partial(
-        build_lora_fused_ops,
-        submodule_names=["up_proj", "down_proj", "gate_proj"],
-        fused_op=KEY_MLP,
-    ),
-    forward_builder_args=["base_type"],
-)
-
-# TODO: have a generic version of this rule
-# - get the module_name and reload on that
-RULE_LLAMA_CE = ModelPatcherRule(
-    rule_id="llama-cross-ent",
-    import_and_maybe_reload=(
-        "torch.nn.CrossEntropyLoss",
-        FastCrossEntropyLoss,
-        "transformers.models.llama.modeling_llama",
-    ),
-)
-
-# TODO: have a generic version of this rule
-# - get the module name
-# - check if "apply_rotary_pos_emb" exists
-# - patch
-RULE_LLAMA_ROPE = ModelPatcherRule(
-    rule_id="llama-rope",
-    import_and_maybe_reload=(
-        "transformers.models.llama.modeling_llama.apply_rotary_pos_emb",
-        fast_rope_embedding,
-        None,
-    ),
-)
-
-LLAMA_MP_RULES = [
-    RULE_LLAMA_RMS,
-    RULE_LLAMA_QKVO,
-    RULE_LLAMA_MLP,
-    RULE_LLAMA_CE,
-    RULE_LLAMA_ROPE,
-]
-
 def get_mp_rules(base_type):
     """
     Function to access all patch rules in this module.
@@ -137,11 +41,92 @@ def get_mp_rules(base_type):
     its forward builder argument, wrap the forward_builder
     function as a partial function with the base_type argument
     """
+    LLAMA_MP_RULES = [
+        # TODO: have a generic version of this rule
+        # - do regex on RMSNorm class name
+        # - check on the tensors required for fast_rms_layernorm
+        ModelPatcherRule(
+            rule_id="llama-rms",
+            trigger=ModelPatcherTrigger(check=LlamaRMSNorm),
+            forward=fast_rms_layernorm,
+        ),
+        # TODO: have a generic version of this rule
+        # - do regex on Attention class name
+        # - have a set of qkv / o module names and check on that
+        ModelPatcherRule(
+            rule_id="llama-qkvo",
+            trigger=combine_triggers(
+                ModelPatcherTrigger(
+                    check=partial(
+                        trigger_fused_ops,
+                        attn_cls=LlamaAttention,
+                        submodule_names=["q_proj", "k_proj", "v_proj"],
+                    )
+                ),
+                ModelPatcherTrigger(
+                    check=partial(
+                        trigger_fused_ops,
+                        attn_cls=LlamaAttention,
+                        submodule_names=["o_proj"],
+                    )
+                ),
+                logic="OR",
+            ),
+            forward_builder=combine_functions(
+                partial(
+                    build_lora_fused_ops,
+                    submodule_names=["q_proj", "k_proj", "v_proj"],
+                    fused_op=KEY_QKV,
+                ),
+                partial(
+                    build_lora_fused_ops,
+                    submodule_names=["o_proj"],
+                    fused_op=KEY_O,
+                ),
+                logic="APPEND",
+            ),
+        ),
+        ModelPatcherRule(
+            rule_id="llama-mlp",
+            trigger=ModelPatcherTrigger(
+                check=partial(
+                    trigger_fused_ops,
+                    attn_cls=LlamaMLP,
+                    submodule_names=["up_proj", "down_proj", "gate_proj"],
+                )
+            ),
+            forward_builder=partial(
+                build_lora_fused_ops,
+                submodule_names=["up_proj", "down_proj", "gate_proj"],
+                fused_op=KEY_MLP,
+            ),
+        ),
+        # TODO: have a generic version of this rule
+        # - get the module_name and reload on that
+        ModelPatcherRule(
+            rule_id="llama-cross-ent",
+            import_and_maybe_reload=(
+                "torch.nn.CrossEntropyLoss",
+                FastCrossEntropyLoss,
+                "transformers.models.llama.modeling_llama",
+            ),
+        ),
+        # TODO: have a generic version of this rule
+        # - get the module name
+        # - check if "apply_rotary_pos_emb" exists
+        # - patch
+        ModelPatcherRule(
+            rule_id="llama-rope",
+            import_and_maybe_reload=(
+                "transformers.models.llama.modeling_llama.apply_rotary_pos_emb",
+                fast_rope_embedding,
+                None,
+            ),
+        )
+    ]
+
     for rule in LLAMA_MP_RULES:
-        if (
-            rule.forward_builder is not None
-            and "base_type" in  rule.forward_builder_args
-        ):
+        if rule.forward_builder is not None:
             rule.forward_builder = partial(
                 rule.forward_builder,
                 base_type=base_type,
