+++ plugins/fused-ops-and-kernels/src/fms_acceleration_foak/models/mistral.py
@@ -35,92 +35,6 @@ from ..kernels.unsloth.rms_layernorm import fast_rms_layernorm
 from ..kernels.unsloth.rope_embedding import fast_rope_embedding
 from .utils import KEY_MLP, KEY_O, KEY_QKV, build_lora_fused_ops, trigger_fused_ops
 
-# - do regex on RMSNorm class name
-# - check on the tensors required for fast_rms_layernorm
-RULE_MISTRAL_RMS = ModelPatcherRule(
-    rule_id="mistral-rms",
-    trigger=ModelPatcherTrigger(check=MistralRMSNorm),
-    forward=fast_rms_layernorm,
-)
-
-RULE_MISTRAL_QKVO = ModelPatcherRule(
-    rule_id="mistral-qkvo",
-    trigger=combine_triggers(
-        ModelPatcherTrigger(
-            check=partial(
-                trigger_fused_ops,
-                attn_cls=MistralAttention,
-                submodule_names=["q_proj", "k_proj", "v_proj"],
-            )
-        ),
-        ModelPatcherTrigger(
-            check=partial(
-                trigger_fused_ops,
-                attn_cls=MistralAttention,
-                submodule_names=["o_proj"],
-            )
-        ),
-        logic="OR",
-    ),
-    forward_builder=combine_functions(
-        partial(
-            build_lora_fused_ops,
-            submodule_names=["q_proj", "k_proj", "v_proj"],
-            fused_op=KEY_QKV,
-        ),
-        partial(
-            build_lora_fused_ops,
-            submodule_names=["o_proj"],
-            fused_op=KEY_O,
-        ),
-        logic="APPEND",
-    ),
-    forward_builder_args=["base_type"],
-)
-
-RULE_MISTRAL_MLP = ModelPatcherRule(
-    rule_id="mistral-mlp",
-    trigger=ModelPatcherTrigger(
-        check=partial(
-            trigger_fused_ops,
-            attn_cls=MistralMLP,
-            submodule_names=["up_proj", "down_proj", "gate_proj"],
-        )
-    ),
-    forward_builder=partial(
-        build_lora_fused_ops,
-        submodule_names=["up_proj", "down_proj", "gate_proj"],
-        fused_op=KEY_MLP,
-    ),
-    forward_builder_args=["base_type"],
-)
-
-RULE_MISTRAL_CE = ModelPatcherRule(
-    rule_id="mistral-cross-ent",
-    import_and_maybe_reload=(
-        "torch.nn.CrossEntropyLoss",
-        FastCrossEntropyLoss,
-        "transformers.models.mistral.modeling_mistral",
-    ),
-)
-
-RULE_MISTRAL_ROPE = ModelPatcherRule(
-    rule_id="mistral-rope",
-    import_and_maybe_reload=(
-        "transformers.models.mistral.modeling_mistral.apply_rotary_pos_emb",
-        fast_rope_embedding,
-        None,
-    ),
-)
-
-MISTRAL_MP_RULES = [
-    RULE_MISTRAL_RMS,
-    RULE_MISTRAL_QKVO,
-    RULE_MISTRAL_MLP,
-    RULE_MISTRAL_CE,
-    RULE_MISTRAL_ROPE,
-]
-
 def get_mp_rules(base_type):
     """
     Function to access all patch rules in this module.
@@ -128,11 +42,83 @@ def get_mp_rules(base_type):
     its forward builder argument, wrap the forward_builder
     function as a partial function with the base_type argument
     """
+
+    MISTRAL_MP_RULES = [
+        # - do regex on RMSNorm class name
+        # - check on the tensors required for fast_rms_layernorm
+        ModelPatcherRule(
+            rule_id="mistral-rms",
+            trigger=ModelPatcherTrigger(check=MistralRMSNorm),
+            forward=fast_rms_layernorm,
+        ),
+        ModelPatcherRule(
+            rule_id="mistral-qkvo",
+            trigger=combine_triggers(
+                ModelPatcherTrigger(
+                    check=partial(
+                        trigger_fused_ops,
+                        attn_cls=MistralAttention,
+                        submodule_names=["q_proj", "k_proj", "v_proj"],
+                    )
+                ),
+                ModelPatcherTrigger(
+                    check=partial(
+                        trigger_fused_ops,
+                        attn_cls=MistralAttention,
+                        submodule_names=["o_proj"],
+                    )
+                ),
+                logic="OR",
+            ),
+            forward_builder=combine_functions(
+                partial(
+                    build_lora_fused_ops,
+                    submodule_names=["q_proj", "k_proj", "v_proj"],
+                    fused_op=KEY_QKV,
+                ),
+                partial(
+                    build_lora_fused_ops,
+                    submodule_names=["o_proj"],
+                    fused_op=KEY_O,
+                ),
+                logic="APPEND",
+            ),
+        ),
+        ModelPatcherRule(
+            rule_id="mistral-mlp",
+            trigger=ModelPatcherTrigger(
+                check=partial(
+                    trigger_fused_ops,
+                    attn_cls=MistralMLP,
+                    submodule_names=["up_proj", "down_proj", "gate_proj"],
+                )
+            ),
+            forward_builder=partial(
+                build_lora_fused_ops,
+                submodule_names=["up_proj", "down_proj", "gate_proj"],
+                fused_op=KEY_MLP,
+            ),
+        ),
+        ModelPatcherRule(
+            rule_id="mistral-cross-ent",
+            import_and_maybe_reload=(
+                "torch.nn.CrossEntropyLoss",
+                FastCrossEntropyLoss,
+                "transformers.models.mistral.modeling_mistral",
+            ),
+        ),
+        ModelPatcherRule(
+            rule_id="mistral-rope",
+            import_and_maybe_reload=(
+                "transformers.models.mistral.modeling_mistral.apply_rotary_pos_emb",
+                fast_rope_embedding,
+                None,
+            ),
+        )
+    ]
+
     for rule in MISTRAL_MP_RULES:
-        if (
-            rule.forward_builder is not None
-            and "base_type" in  rule.forward_builder_args
-        ):
+        if rule.forward_builder is not None:
             rule.forward_builder = partial(
                 rule.forward_builder,
                 base_type=base_type,
