+++ plugins/fused-ops-and-kernels/src/fms_acceleration_foak/models/mixtral.py
@@ -34,74 +34,6 @@ from ..kernels.unsloth.rope_embedding import fast_rope_embedding
 
 from .utils import KEY_O, KEY_QKV, build_lora_fused_ops, trigger_fused_ops
 
-# - do regex on RMSNorm class name
-# - check on the tensors required for fast_rms_layernorm
-RULE_MIXTRAL_RMS = ModelPatcherRule(
-    rule_id="mixtral-rms",
-    trigger=ModelPatcherTrigger(check=MixtralRMSNorm),
-    forward=fast_rms_layernorm,
-)
-
-RULE_MIXTRAL_QKVO = ModelPatcherRule(
-    rule_id="mixtral-qkvo",
-    trigger=combine_triggers(
-        ModelPatcherTrigger(
-            check=partial(
-                trigger_fused_ops,
-                attn_cls=MixtralAttention,
-                submodule_names=["q_proj", "k_proj", "v_proj"],
-            )
-        ),
-        ModelPatcherTrigger(
-            check=partial(
-                trigger_fused_ops,
-                attn_cls=MixtralAttention,
-                submodule_names=["o_proj"],
-            )
-        ),
-        logic="OR",
-    ),
-    forward_builder=combine_functions(
-        partial(
-            build_lora_fused_ops,
-            submodule_names=["q_proj", "k_proj", "v_proj"],
-            fused_op=KEY_QKV,
-        ),
-        partial(
-            build_lora_fused_ops,
-            submodule_names=["o_proj"],
-            fused_op=KEY_O,
-        ),
-        logic="APPEND",
-    ),
-    forward_builder_args=["base_type"],
-)
-
-RULE_MIXTRAL_CE = ModelPatcherRule(
-    rule_id="mixtral-cross-ent",
-    import_and_maybe_reload=(
-        "torch.nn.CrossEntropyLoss",
-        FastCrossEntropyLoss,
-        "transformers.models.mixtral.modeling_mixtral",
-    ),
-)
-
-RULE_MIXTRAL_ROPE = ModelPatcherRule(
-    rule_id="mixtral-rope",
-    import_and_maybe_reload=(
-        "transformers.models.mixtral.modeling_mixtral.apply_rotary_pos_emb",
-        fast_rope_embedding,
-        None,
-    ),
-)
-
-MIXTRAL_MP_RULES = [
-    RULE_MIXTRAL_RMS,
-    RULE_MIXTRAL_QKVO,
-    RULE_MIXTRAL_CE,
-    RULE_MIXTRAL_ROPE,
-]
-
 def get_mp_rules(base_type):
     """
     Function to access all patch rules in this module.
@@ -109,11 +41,68 @@ def get_mp_rules(base_type):
     its forward builder argument, wrap the forward_builder
     function as a partial function with the base_type argument
     """
+
+    # - do regex on RMSNorm class name
+    # - check on the tensors required for fast_rms_layernorm
+    MIXTRAL_MP_RULES = [
+        ModelPatcherRule(
+            rule_id="mixtral-rms",
+            trigger=ModelPatcherTrigger(check=MixtralRMSNorm),
+            forward=fast_rms_layernorm,
+        ),
+        ModelPatcherRule(
+            rule_id="mixtral-qkvo",
+            trigger=combine_triggers(
+                ModelPatcherTrigger(
+                    check=partial(
+                        trigger_fused_ops,
+                        attn_cls=MixtralAttention,
+                        submodule_names=["q_proj", "k_proj", "v_proj"],
+                    )
+                ),
+                ModelPatcherTrigger(
+                    check=partial(
+                        trigger_fused_ops,
+                        attn_cls=MixtralAttention,
+                        submodule_names=["o_proj"],
+                    )
+                ),
+                logic="OR",
+            ),
+            forward_builder=combine_functions(
+                partial(
+                    build_lora_fused_ops,
+                    submodule_names=["q_proj", "k_proj", "v_proj"],
+                    fused_op=KEY_QKV,
+                ),
+                partial(
+                    build_lora_fused_ops,
+                    submodule_names=["o_proj"],
+                    fused_op=KEY_O,
+                ),
+                logic="APPEND",
+            ),
+        ),
+        ModelPatcherRule(
+            rule_id="mixtral-cross-ent",
+            import_and_maybe_reload=(
+                "torch.nn.CrossEntropyLoss",
+                FastCrossEntropyLoss,
+                "transformers.models.mixtral.modeling_mixtral",
+            ),
+        ),
+        ModelPatcherRule(
+            rule_id="mixtral-rope",
+            import_and_maybe_reload=(
+                "transformers.models.mixtral.modeling_mixtral.apply_rotary_pos_emb",
+                fast_rope_embedding,
+                None,
+            ),
+        )
+    ]
+
     for rule in MIXTRAL_MP_RULES:
-        if (
-            rule.forward_builder is not None
-            and "base_type" in  rule.forward_builder_args
-        ):
+        if rule.forward_builder is not None:
             rule.forward_builder = partial(
                 rule.forward_builder,
                 base_type=base_type,
